+++ sema/type_checker.h
@@ -40,6 +40,7 @@ public:
     void visit(BlockStmt* node) override;
     void visit(IfStmt* node) override;
     void visit(WhileStmt* node) override;
+    void visit(DoWhileStmt* node) override;
     void visit(ForStmt* node) override;
     void visit(ForInStmt* node) override;
     void visit(ReturnStmt* node) override;
@@ -48,7 +49,9 @@ public:
 
     void visit(BinaryExpr* node) override;
     void visit(UnaryExpr* node) override;
+    void visit(IncDecExpr* node) override;
     void visit(QuestionExpr* node) override;
+    void visit(TernaryExpr* node) override;
     void visit(CallExpr* node) override;
     void visit(IndexExpr* node) override;
     void visit(MemberExpr* node) override;
@@ -56,6 +59,7 @@ public:
     void visit(LiteralExpr* node) override;
     void visit(IdentExpr* node) override;
     void visit(StructInitExpr* node) override;
+    void visit(ArrayLitExpr* node) override;
     void visit(AllocWithExpr* node) override;
     void visit(LambdaExpr* node) override;
     void visit(AsmStmt* node) override;